feat(db): v1.2.3 - Decimal precision for monetary fields
BREAKING CHANGE: Monetary columns migrated from DOUBLE PRECISION to NUMERIC

- events.price → NUMERIC(10,2)
- service_types.price → NUMERIC(10,2)
- bookings.amount_paid → NUMERIC(10,2)
- ai_usage_logs.cost_provider_usd → NUMERIC(10,6)
- ai_usage_logs.cost_user_credits → NUMERIC(10,4)

Migration: b7573685053a_refactor_monetary_fields_to_decimal.py
Pydantic schemas updated to use Decimal type

+++ backend/app/api/v1/booking.py
@@ -17,6 +17,7 @@ router = APIRouter()
 
 
 # ============ SCHEMAS ============
+from decimal import Decimal
 
 
 class BookingListResponse(BaseModel):
@@ -27,11 +28,11 @@ class BookingListResponse(BaseModel):
     patient_name: str
     service_id: uuid.UUID
     service_title: str
-    service_price: float  # For pending payment calculations
+    service_price: Decimal  # For pending payment calculations
     start_time: datetime
     end_time: datetime
     status: str
-    amount_paid: float
+    amount_paid: Decimal
     currency: str
     created_at: datetime
